v2

+++ app/Http/Controllers/Admin/ChestController.php
@@ -10,11 +10,15 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\DB;
 use App\User;
+use Illuminate\Support\Facades\Gate;
 
 class ChestController extends Controller
 {
   public function store(Request $request)
   {
+    if (!Gate::allows('admin')) {
+      return $this->sendError('Forbidden', [], 403);
+    }
     Cache::forget('chests');
     $chest = Chest::create($request->all());
     $chest->saveItems(json_decode($request->get('items')));
@@ -25,6 +29,9 @@ class ChestController extends Controller
 
   public function index()
   {
+    if (!Gate::allows('admin')) {
+      return $this->sendError('Forbidden', [], 403);
+    }
     $chests = Chest::with('items')->get();
 
     return $this->sendResponse(ChestResource::collection($chests), 'Ok', 200);
@@ -32,6 +39,9 @@ class ChestController extends Controller
 
   public function deleteById(Request $request, $id)
   {
+    if (!Gate::allows('admin')) {
+      return $this->sendError('Forbidden', [], 403);
+    }
     Cache::forget('chests');
     $result = Chest::find($id)->delete();
 
@@ -40,6 +50,9 @@ class ChestController extends Controller
 
   public function update(Request $request, $id)
   {
+    if (!Gate::allows('admin')) {
+      return $this->sendError('Forbidden', [], 403);
+    }
     Cache::forget('chests');
     $chest = Chest::find($id);
     $chest->update($request->all());
@@ -54,6 +67,9 @@ class ChestController extends Controller
 
   public function getById($id)
   {
+    if (!Gate::allows('admin')) {
+      return $this->sendError('Forbidden', [], 403);
+    }
     $chest = Chest::with('items')->find($id);
 
     return $this->sendResponse(new ChestResource($chest), 'Ok', 200);
